introduce ConfigurationBuilder
