+++ src/GitVersionCore.Tests/Extensions/GitToolsTestingExtensions.cs
@@ -21,7 +21,7 @@ namespace GitVersionCore.Tests
         {
             if (configuration == null)
             {
-                configuration = DefaultConfigProvider.CreateDefaultConfig();
+                configuration = new ConfigurationBuilder().Build();
             }
 
             repository ??= fixture.Repository;
@@ -67,7 +67,7 @@ namespace GitVersionCore.Tests
         public static void AssertFullSemver(this RepositoryFixtureBase fixture, string fullSemver, Config configuration = null, IRepository repository = null, string commitId = null, bool onlyTrackedBranches = true, string targetBranch = null)
         {
             configuration ??= new Config();
-            configuration = DefaultConfigProvider.CreateDefaultConfig().Apply(configuration).FinalizeConfig();
+            configuration = new ConfigurationBuilder().Add(configuration).Build();
             Console.WriteLine("---------");
 
             try
