+++ src/GitVersionCore.Tests/Helpers/TestBase.cs
@@ -26,7 +26,7 @@ namespace GitVersionCore.Tests.Helpers
 
         protected static IServiceProvider BuildServiceProvider(string workingDirectory, IRepository repository, string branch, Config config = null)
         {
-            config ??= DefaultConfigProvider.CreateDefaultConfig().FinalizeConfig();
+            config ??= new ConfigurationBuilder().Build();
             var options = Options.Create(new GitVersionOptions
             {
                 WorkingDirectory = workingDirectory,
