+++ src/GitVersionCore.Tests/IntegrationTests/VersionInTagScenarios.cs
@@ -14,12 +14,12 @@ namespace GitVersionCore.Tests.IntegrationTests
         public void TagPreReleaseWeightIsNotConfigured_HeadIsATaggedCommit_WeightedPreReleaseNumberShouldBeTheDefaultValue()
         {
             // Arrange
-            var config = DefaultConfigProvider.CreateDefaultConfig()
-                                              .Apply(new Config
-                                              {
-                                                  AssemblyFileVersioningFormat = "{Major}.{Minor}.{Patch}.{WeightedPreReleaseNumber}",
-                                              })
-                                              .FinalizeConfig();
+            var config = new ConfigurationBuilder()
+                         .Add(new Config
+                         {
+                             AssemblyFileVersioningFormat = "{Major}.{Minor}.{Patch}.{WeightedPreReleaseNumber}",
+                         })
+                         .Build();
 
             // Act
             using var fixture = new BaseGitFlowRepositoryFixture("1.0.0");
@@ -34,13 +34,13 @@ namespace GitVersionCore.Tests.IntegrationTests
         public void TagPreReleaseWeightIsConfigured_HeadIsATaggedCommit_WeightedPreReleaseNumberShouldBeTheSameAsTheTagPreReleaseWeight()
         {
             // Arrange
-            var config = DefaultConfigProvider.CreateDefaultConfig()
-                                              .Apply(new Config
-                                              {
-                                                  AssemblyFileVersioningFormat = "{Major}.{Minor}.{Patch}.{WeightedPreReleaseNumber}",
-                                                  TagPreReleaseWeight = 65535
-                                              })
-                                              .FinalizeConfig();
+            var config = new ConfigurationBuilder()
+                         .Add(new Config
+                         {
+                             AssemblyFileVersioningFormat = "{Major}.{Minor}.{Patch}.{WeightedPreReleaseNumber}",
+                             TagPreReleaseWeight = 65535
+                         })
+                         .Build();
 
             // Act
             using var fixture = new BaseGitFlowRepositoryFixture("1.0.0");
@@ -55,14 +55,14 @@ namespace GitVersionCore.Tests.IntegrationTests
         public void TagPreReleaseWeightIsConfigured_GitFlowReleaseIsFinished_WeightedPreReleaseNumberShouldBeTheSameAsTheTagPreReleaseWeight()
         {
             // Arrange
-            var config = DefaultConfigProvider.CreateDefaultConfig()
-                                              .Apply(new Config
-                                              {
-                                                  AssemblyFileVersioningFormat = "{Major}.{Minor}.{Patch}.{WeightedPreReleaseNumber}",
-                                                  TagPreReleaseWeight = 65535,
-                                                  VersioningMode = VersioningMode.ContinuousDeployment
-                                              })
-                                              .FinalizeConfig();
+            var config = new ConfigurationBuilder()
+                         .Add(new Config
+                         {
+                             AssemblyFileVersioningFormat = "{Major}.{Minor}.{Patch}.{WeightedPreReleaseNumber}",
+                             TagPreReleaseWeight = 65535,
+                             VersioningMode = VersioningMode.ContinuousDeployment
+                         })
+                         .Build();
 
             // Act
             using var fixture = new BaseGitFlowRepositoryFixture("1.0.0");
@@ -84,13 +84,13 @@ namespace GitVersionCore.Tests.IntegrationTests
         public void TagPreReleaseWeightIsNotConfigured_GitFlowReleaseIsFinished_WeightedPreReleaseNumberShouldBeTheDefaultValue()
         {
             // Arrange
-            var config = DefaultConfigProvider.CreateDefaultConfig()
-                                              .Apply(new Config
-                                              {
-                                                  AssemblyFileVersioningFormat = "{Major}.{Minor}.{Patch}.{WeightedPreReleaseNumber}",
-                                                  VersioningMode = VersioningMode.ContinuousDeployment
-                                              })
-                                              .FinalizeConfig();
+            var config = new ConfigurationBuilder()
+                         .Add(new Config
+                         {
+                             AssemblyFileVersioningFormat = "{Major}.{Minor}.{Patch}.{WeightedPreReleaseNumber}",
+                             VersioningMode = VersioningMode.ContinuousDeployment
+                         })
+                         .Build();
 
             // Act
             using var fixture = new BaseGitFlowRepositoryFixture("1.0.0");
